% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ATCodeR.R
\name{ATCtransform}
\alias{ATCtransform}
\title{Transform free-text substances into ATC Codes}
\usage{
ATCtransform(input.df, column_name)
}
\arguments{
\item{column_name}{The name of the column in `input_df` that contains free-text substances.}

\item{input_df}{A data frame containing the column specified by `column_name`.}
}
\value{
A data frame with the transformed data(focus on antineoplastic medication) and additional columns.
}
\description{
Transform free-text substances into ATC Codes
}
